Extract Parser interface and declare CsvFileParser
